requirements.txt
torch
torchvision
numpy
matplotlib
avalanche-lib
transformers